Fixed warnings/errors wth

+++ SDWebImage/SDWebImageDownloaderOperation.m
@@ -228,7 +228,7 @@ static NSURLCredential* sCurrentCreds = nil;
         // Thanks to the author @Nyx0uf
 
         // Get the total bytes downloaded
-        const NSUInteger totalSize = self.imageData.length;
+        const long long totalSize = (long long)self.imageData.length;
 
         // Update the data source, we must pass ALL the data, not just the new bytes
         CGImageSourceRef imageSource = CGImageSourceCreateIncremental(NULL);
